requirements.txt
pyyaml
discord.py
python-dotenv
requests
pytest
codecov
coverage
pytest-cov